Added a Delta Time Functionality
Also added a Platform static class for generic platform specific
operations like getting time and such. Some #include's have also been
ammended to work properly in MainGame

+++ Engine/src/Oyl3D/Graphics/Buffer.cpp
@@ -6,7 +6,7 @@
 
 namespace oyl {
 
-oyl::VertexBuffer* VertexBuffer::create(float* vertices, size_t size) {
+oyl::VertexBuffer* VertexBuffer::create(float* vertices, uint size) {
 	switch (Renderer::getAPI()) {
 	case RendererAPI::API::None: ASSERT(false, "None is currently unsupported"); return nullptr;
 	case RendererAPI::API::OpenGL: return new OpenGLVertexBuffer(vertices, size);
@@ -14,7 +14,7 @@ oyl::VertexBuffer* VertexBuffer::create(float* vertices, size_t size) {
 	return nullptr;
 }
 
-oyl::IndexBuffer* IndexBuffer::create(uint* indices, size_t size) {
+oyl::IndexBuffer* IndexBuffer::create(uint* indices, uint size) {
 	switch (Renderer::getAPI()) {
 	case RendererAPI::API::None: ASSERT(false, "None is currently unsupported"); return nullptr;
 	case RendererAPI::API::OpenGL: return new OpenGLIndexBuffer(indices, size);
